Moving right along. Embedding different kind of blanket replacements, which is what the new Alice and Massage classes are. Onward and upward!

+++ Program.cs
@@ -6,7 +6,8 @@ namespace MetX.VB6ToCSharp
     public static class Program
     {
         public const string SourceCodeFolderPath = @"I:\OneDrive\data\code\Slice and Dice\Sandy\";
-        public const string OutputFolderPath = @"I:\OneDrive\data\code\Slice and Dice\SandyB\";
+        public const string OutputFolderPath = @"I:\OneDrive\data\code\Slice and Dice\SandyC\";
+        public const bool ClearOutputFolder = false;
 
         public static int Main(string[] args)
         {
@@ -16,12 +17,13 @@ namespace MetX.VB6ToCSharp
 
         public static void ConvertAllFiles()
         {
-            // Delete previous run
-            foreach (var fileToDelete in Directory.EnumerateFiles(OutputFolderPath))
-            {
-                File.SetAttributes(fileToDelete, FileAttributes.Normal);
-                File.Delete(fileToDelete);
-            }
+            if(ClearOutputFolder)
+                // Delete previous run
+                foreach (var fileToDelete in Directory.EnumerateFiles(OutputFolderPath))
+                {
+                    File.SetAttributes(fileToDelete, FileAttributes.Normal);
+                    File.Delete(fileToDelete);
+                }
 
             var fileSets = new List<IEnumerable<string>>
             {
